requirements_gpu.txt
polars 
pandas 
matplotlib 
google-cloud-bigquery
google-cloud-storage
biopython 
networkx
torch --index-url https://download.pytorch.org/whl/cu118  
torchvision --index-url https://download.pytorch.org/whl/cu118  
torchaudio --index-url https://download.pytorch.org/whl/cu118 
transformers
torch_geometric